fixed wrong leader position calculation
In line 64, We set the Leader position equals the position of current whale. This passes the variable "Positions[i,:]" by reference and hence the "leader_pos" is changed in the next loop. So, we need to copy - by value not by reference - the position of the current whale into the "Leader_Pos".