fixed wrong calculation of leaders positions
In lines 60, 65 and 70. The leader's position is assigned the  valued of "Positions[i,:]". This actually passes the variable "Positions[i,:]" by reference and hence the leader position is changed in the next loop even if the current solution does not have a lower/higher fitness.
So, we need to copy - by value not by reference - the current position vector into the leaders variables.